fix KT-5866. Moved classes into inlines package
 Also copy-pasted static-initializer of asm.Frame.OPCODES
 because it package-protected
