+++ compiler/backend/src/org/jetbrains/jet/codegen/inline/MaxLocalsCalculator.java
@@ -14,7 +14,12 @@
  * limitations under the License.
  */
 
-package org.jetbrains.org.objectweb.asm;
+package org.jetbrains.jet.codegen.inline;
+
+import org.jetbrains.org.objectweb.asm.Label;
+import org.jetbrains.org.objectweb.asm.MethodVisitor;
+import org.jetbrains.org.objectweb.asm.Opcodes;
+import org.jetbrains.org.objectweb.asm.Type;
 
 public class MaxLocalsCalculator extends MethodVisitor {
